extract filter functions to FilterService class
